+++ opteryx/engine/functions/date_functions.py
@@ -9,48 +9,55 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+import datetime
+import numpy
 
 from pyarrow import compute
 
-import datetime
-import numpy
 from opteryx.exceptions import SqlError
-
 from opteryx.utils.dates import parse_iso
 
 
 def get_time():
-    # doing this inline means the utcnow() function is called once and time is then
-    # called for each row - meaning it's not 'now', it's 'when did you start'
+    """
+    Get the current time
+    """
     return datetime.datetime.utcnow().time()
 
 
 def get_yesterday():
-    return datetime.date.today() - datetime.timedelta(days=1)
+    """
+    calculate yesterday
+    """
+    return datetime.datetime.utcnow().date() - datetime.timedelta(days=1)
 
 
-def get_date(input):
+def get_date(timestamp):
     """
     Convert input to a datetime object and extract the Date part
     """
     # if it's a string, parse it (to a datetime)
-    if isinstance(input, str):
-        input = parse_iso(input)
+    if isinstance(timestamp, str):
+        timestamp = parse_iso(timestamp)
     # if it's a numpy datetime, convert it to a date
-    if isinstance(input, (numpy.datetime64)):
-        input = input.astype("M8[D]").astype("O")
+    if isinstance(timestamp, (numpy.datetime64)):
+        timestamp = timestamp.astype("M8[D]").astype("O")
     # if it's a datetime, convert it to a date
-    if isinstance(input, datetime.datetime):
-        input = input.date()
+    if isinstance(timestamp, datetime.datetime):
+        timestamp = timestamp.date()
     # set it to midnight that day to make it a datetime
-    if isinstance(input, datetime.date):
-        return datetime.datetime.combine(input, datetime.datetime.min.time())
+    # even though we're getting the date, the supported column type is datetime
+    if isinstance(timestamp, datetime.date):
+        return datetime.datetime.combine(timestamp, datetime.datetime.min.time())
     return None
 
 
 def date_part(interval, arr):
+    """
+    Also the EXTRACT function - we extract a given part from an array of dates
+    """
 
-    EXTRACTOR = {
+    extractors = {
         "microsecond": compute.microsecond,
         "second": compute.second,
         "minute": compute.minute,
@@ -65,7 +72,7 @@ def date_part(interval, arr):
     }
 
     interval = interval.lower()
-    if interval in EXTRACTOR:
-        return EXTRACTOR[interval](arr)
+    if interval in extractors:
+        return extractors[interval](arr)
 
-    raise SqlError(f"Unsupported date part `{interval}`")
+    raise SqlError(f"Unsupported date part `{interval}`")  # pragma: no cover