+++ opteryx/engine/planner/operations/internal_dataset_node.py
@@ -56,7 +56,7 @@ class InternalDatasetNode(BasePlanNode):
         The Blob Reader Node is responsible for reading the relevant blobs
         and returning a Table/Relation.
         """
-        super().__init__(directives=directives, statistics=statistics, **config)
+        super().__init__(directives=directives, statistics=statistics)
 
         self._statistics = statistics
         self._alias = config["alias"]